Create skeleton of project
The skeleton of this project was created. With that the next steps
are allowed to start. This is the basic skeleton created by Rails.

The next steps are to configurate to utilize Docker and the Gems that will be use
during this process of development.

As a result, the skeleton of this project is ready for start the process to development