Configure project to utilize Docker
Configure the project to utilize Docker with Docker compose. There are created more two files start and start_bjos
who your responsiblite it start command during execute docker commands. The start file will check if the gems was installed and case not will run bundle install,after that
it will start puma server. The start_jobs will run the command to initialize the sidekiq server and create a new queue call emails

As a result, the project is ready to utilize Docker and Docker compose to start server and run application